Move Point class to its own file
